JPERF-715: Extract `WaitingCommand` for reuse
